set preferred port for networkserialport
This allows for specifying a start port and a range up from that to be
used when looking for a free port to use with ser2net when creating a
networkserialport resource. This can be helpful on networks where access
to ports on the network is limited.

Signed-off-by: Benjamin B. Frost <benjamin@geanix.com>
(cherry picked from commit 85b24137af8b684fe230ac03739b8035c0dae5ea)